SoulCompass/requirements.txt
streamlit>=1.43.0
pandas>=2.0.0
pillow>=9.0.0
